03-a2a-agent/requirements.txt
azure-ai-projects[agents]
azure-identity
python-dotenv
